+++ antarest/core/configdata/repository.py
@@ -16,7 +16,7 @@ from typing import Optional
 from antarest.core.configdata.model import ConfigData
 from antarest.core.jwt import DEFAULT_ADMIN_USER
 from antarest.core.model import JSON
-from antarest.core.serialization.utils import from_json, to_json
+from antarest.core.serialization import from_json, to_json_string
 from antarest.core.utils.fastapi_sqlalchemy import db
 
 
@@ -50,7 +50,7 @@ class ConfigDataRepository:
     def put_json(self, key: str, data: JSON, owner: Optional[int] = None) -> None:
         configdata = ConfigData(
             key=key,
-            value=to_json(data).decode("utf-8"),
+            value=to_json_string(data),
             owner=owner or DEFAULT_ADMIN_USER.id,
         )
         configdata = db.session.merge(configdata)
