+++ antarest/eventbus/web.py
@@ -26,7 +26,7 @@ from antarest.core.interfaces.eventbus import Event, IEventBus
 from antarest.core.jwt import DEFAULT_ADMIN_USER, JWTUser
 from antarest.core.model import PermissionInfo, StudyPermissionType
 from antarest.core.permissions import check_permission
-from antarest.core.serialization.utils import to_json
+from antarest.core.serialization import to_json_string
 from antarest.fastapi_jwt_auth import AuthJWT
 from antarest.login.auth import Auth
 
@@ -99,7 +99,7 @@ def configure_websockets(app_ctxt: AppBuildContext, config: Config, event_bus: I
         event_data = event.model_dump()
         del event_data["permissions"]
         del event_data["channel"]
-        await manager.broadcast(to_json(event_data).decode("utf-8"), event.permissions, event.channel)
+        await manager.broadcast(to_json_string(event_data), event.permissions, event.channel)
 
     @app_ctxt.api_root.websocket("/ws")
     async def connect(
