+++ antarest/login/auth.py
@@ -21,7 +21,7 @@ from starlette.requests import Request
 
 from antarest.core.config import Config
 from antarest.core.jwt import DEFAULT_ADMIN_USER, JWTUser
-from antarest.core.serialization.utils import from_json
+from antarest.core.serialization import from_json
 from antarest.fastapi_jwt_auth import AuthJWT
 
 logger = logging.getLogger(__name__)
