+++ antarest/study/storage/patch_service.py
@@ -13,7 +13,7 @@
 import typing as t
 from pathlib import Path
 
-from antarest.core.serialization.utils import from_json
+from antarest.core.serialization import from_json
 from antarest.study.model import Patch, PatchOutputs, RawStudy, StudyAdditionalData
 from antarest.study.repository import StudyMetadataRepository
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
