+++ antarest/study/storage/rawstudy/model/filesystem/config/files.py
@@ -21,7 +21,7 @@ from enum import Enum
 from pathlib import Path
 
 from antarest.core.model import JSON
-from antarest.core.serialization.utils import from_json
+from antarest.core.serialization import from_json
 from antarest.study.storage.rawstudy.ini_reader import IniReader
 from antarest.study.storage.rawstudy.model.filesystem.config.binding_constraint import (
     DEFAULT_GROUP,
