+++ antarest/study/storage/rawstudy/model/filesystem/config/ini_properties.py
@@ -14,7 +14,7 @@ import typing as t
 
 from pydantic import BaseModel
 
-from antarest.core.serialization.utils import from_json, to_json
+from antarest.core.serialization import from_json, to_json
 
 
 class IniProperties(
