+++ antarest/study/web/xpansion_studies_blueprint.py
@@ -20,7 +20,7 @@ from antarest.core.config import Config
 from antarest.core.jwt import JWTUser
 from antarest.core.model import JSON, StudyPermissionType
 from antarest.core.requests import RequestParameters
-from antarest.core.serialization.utils import to_json
+from antarest.core.serialization import to_json
 from antarest.core.utils.web import APITag
 from antarest.login.auth import Auth
 from antarest.study.business.xpansion_management import (
