+++ antarest/tools/lib.py
@@ -24,7 +24,7 @@ from httpx import Client
 
 from antarest.core.cache.business.local_chache import LocalCache
 from antarest.core.config import CacheConfig
-from antarest.core.serialization.utils import from_json, to_json_as_bytes
+from antarest.core.serialization import from_json, to_json_string
 from antarest.core.tasks.model import TaskDTO
 from antarest.core.utils.utils import StopWatch, get_local_path
 from antarest.matrixstore.repository import MatrixContentRepository
@@ -208,7 +208,7 @@ def extract_commands(study_path: Path, commands_output_dir: Path) -> None:
     command_list = extractor.extract(study)
 
     (commands_output_dir / COMMAND_FILE).write_text(
-        to_json_as_bytes([command.model_dump(exclude={"id"}) for command in command_list], indent=2)
+        to_json_string([command.model_dump(exclude={"id"}) for command in command_list], indent=2)
     )
 
 
@@ -297,7 +297,7 @@ def generate_diff(
     )
 
     (output_dir / COMMAND_FILE).write_text(
-        to_json_as_bytes([command.to_dto().model_dump(exclude={"id"}) for command in diff_commands], indent=2)
+        to_json_string([command.to_dto().model_dump(exclude={"id"}) for command in diff_commands], indent=2)
     )
 
     needed_matrices: Set[str] = set()
